pyproject.toml
[tool.poetry]
name = "discord_py_bot"
version = "0.1.0"
description = ""
authors = ["Skyler Burger <skylerburger@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
python-environ = "^0.4.54"
google-api-python-client = "^2.106.0"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
pytest-asyncio = "^0.14.0"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
